+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## next
+  - Add support for `cxx-options` and `cxx-sources` (see #205)
+
 ## Changes in 0.27.0
   - Local defaults are now resolved relative to the file they are
     mentioned in, not the CWD that hpack is invoked from.